[bitnami/elasticsearch] attach PVC's for es master node data
Elasticsearch master nodes store the cluster status at `/bitnami/elasticsearch/data`. Among other things
this includes the UUID of the elasticsearch cluster. Without a persistent data store for this data, the
UUID of a cluster could change if k8s node(s) hosting the es master nodes go down and are scheduled on
some other master node. In the event that this happens, the data nodes will no longer be able to join
a cluster as the uuid changed resulting in a broken cluster.

To resolve such issues, this PR attaches PVC's to the master nodes data store.

Signed-off-by: Sameer Naik <sameersbn@vmware.com>